core: don't check for writen output to terminate
commit 9e0bcaad2 and 744def043e are thought to fix the decryption of a
short artifact of one AES block. However, check is done on both input
and output data, but just checking if there are still data in the
incoming buffer is enough. This allows a crypto module to postpone the
decryption and to collect data - some algorithms are not able to work on
stream of data, and they need to have the full crypto message in memory.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>